+++ src/main/java/com/bankcli/service/AccountService.java
@@ -1,14 +1,15 @@
 package com.bankcli.service;
 
 import com.bankcli.domain.Account;
+import java.math.BigDecimal;
 
 public interface AccountService {
 	void getAccount(Account account);
 	void getAccountStatus();
-	void makeDeposit(double amount);
-	void makeWithdrawal(double amount);
+	void makeDeposit(BigDecimal amount);
+	void makeWithdrawal(BigDecimal amount);
 	void createAccount(String pin);
-	void makeTransfer(int toAccountId, double amount);
+	void makeTransfer(int toAccountId, BigDecimal amount);
 	void logOut();
 	void getTransactions();
 }
